make test directory, Separate the handleCalculate function from App.jsx and update the run test command in package.json
