+++ frontend/package.json
@@ -9,7 +9,7 @@
     "lint": "eslint src --ext js,jsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview",
     "start": "react-scripts start",
-    "test": "mocha",
+    "test": "mocha --require babel-register ./src/test/App.test.js",
     "eject": "react-scripts eject"
   },
   "dependencies": {
