+++ frontend/src/App.jsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from "react";
 import logger from "./logger";
 import "./App.css";
 import axios from "axios";
-
+import { handleCalculate } from "./hooks/handleCalculate";
 const App = () => {
   const [valueOne, setValueOne] = useState(0);
   const [valueTwo, setValueTwo] = useState(0);
@@ -23,26 +23,12 @@ const App = () => {
   };
 
   //caluclate based on the operator
-  const handleCalculate = () => {
-    let calculatedResult;
-    switch (selectedOperation) {
-      case "+":
-        calculatedResult = parseInt(operandOne) + parseInt(operandTwo);
-        break;
-      case "-":
-        calculatedResult = parseInt(operandOne) - parseInt(operandTwo);
-        break;
-      case "*":
-        calculatedResult = parseInt(operandOne) * parseInt(operandTwo);
-        break;
-      case "/":
-        calculatedResult = parseInt(operandOne) / parseInt(operandTwo);
-        break;
-      default:
-        calculatedResult = 0;
-    }
-    setResult(calculatedResult);
-    handleSubmit(selectedOperation, calculatedResult);
+  const calculate = () => {
+    setResult(handleCalculate(selectedOperation, operandOne, operandTwo));
+    handleSubmit(
+      selectedOperation,
+      handleCalculate(selectedOperation, operandOne, operandTwo)
+    );
   };
 
   //handling submission
@@ -107,6 +93,7 @@ const App = () => {
                 setValueOne(e.target.value);
                 setOperandOne(e.target.value);
               }}
+              onInput={() => setResult(0)}
             />
           </div>
           <div className="button-container">
@@ -136,7 +123,7 @@ const App = () => {
           </div>
         </div>
         <div className="submit-container">
-          <button name="=" onClick={handleCalculate} className="submit-button">
+          <button name="=" onClick={calculate} className="submit-button">
             =
           </button>
         </div>
