+++ frontend/src/hooks/handleCalculate.js
